:memo: Rename airtable_data to Assets
